pyproject.toml
[project]
name = "prettyqr"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pillow>=11.2.1", "qrcode>=8.2"]
